Adjust option parameter handling
We allow optional command line arguments, but in the end we want a
configuration struct that has concrete values for everything. Hence,
create a second struct without Option<T> members and have
sanitize_arguments() return an instance of that.

While we're here, add some docs for parameters and a `--port` argument.